// examples/blog/types/database.ts
/**
 * Generated by: supatype generate
 * Do not edit manually — regenerate with: npx supatype generate
 */

export interface Database {
  public: {
    Tables: {
      users: {
        Row: {
          id: string
          name: string
          avatar_url: string | null
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: string | undefined
          name: string
          avatar_url?: string | null | undefined
          created_at?: string | undefined
          updated_at?: string | undefined
        }
        Update: {
          id?: string | undefined
          name?: string | undefined
          avatar_url?: string | null | undefined
          updated_at?: string | undefined
        }
      }
      posts: {
        Row: {
          id: string
          title: string
          slug: string
          body: string
          cover_image: string | null
          author_id: string
          status: "draft" | "published" | "archived"
          published_at: string | null
          scheduled_at: string | null
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: string | undefined
          title: string
          slug?: string | undefined
          body: string
          cover_image?: string | null | undefined
          author_id: string
          status?: "draft" | "published" | "archived" | undefined
          published_at?: string | null | undefined
          scheduled_at?: string | null | undefined
          created_at?: string | undefined
          updated_at?: string | undefined
        }
        Update: {
          title?: string | undefined
          slug?: string | undefined
          body?: string | undefined
          cover_image?: string | null | undefined
          author_id?: string | undefined
          status?: "draft" | "published" | "archived" | undefined
          published_at?: string | null | undefined
          scheduled_at?: string | null | undefined
          updated_at?: string | undefined
        }
      }
      comments: {
        Row: {
          id: string
          post_id: string
          author_id: string
          body: string
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: string | undefined
          post_id: string
          author_id: string
          body: string
          created_at?: string | undefined
          updated_at?: string | undefined
        }
        Update: {
          post_id?: string | undefined
          author_id?: string | undefined
          body?: string | undefined
          updated_at?: string | undefined
        }
      }
    }
  }
}